Complete comprehensive warning elimination and context transfer preparation
- ✅ Achieved clean ARM GCC compilation (52,180 bytes text, 89,636 bytes total)
- ✅ Eliminated all user-controllable compilation warnings across motor control modules
- ✅ Updated STATUS.md with comprehensive context transfer documentation
- ✅ Enhanced semantic search system with 77,938 documents indexed
- ✅ Validated development environment and build system operational
- 🔧 Added optimization telemetry system and motor efficiency optimization
- 🔧 Enhanced mock frameworks and unit testing infrastructure
- 📋 Ready for hardware validation phase with STM32H753ZI + X-NUCLEO-IHM02A1

Key modules cleaned:
- motion_profile.c/h: Static function organization
- position_control.c/h: Forward declarations and signatures
- multi_motor_coordinator.c/h: Function visibility fixes
- real_time_control.c: Compilation warning elimination
- communication modules: Protocol and configuration updates

Next: Hardware testing and AS5600 encoder integration validation

+++ Core/Src/main.c
@@ -204,8 +204,9 @@ int main(void) {
     printf("  - Telemetry: Priority %d, %d words stack, %d ms period\r\n",
            TELEMETRY_TASK_PRIORITY, TELEMETRY_TASK_STACK_SIZE,
            TELEMETRY_PERIOD_MS);
-    printf("  - Default Task: Priority %d (lowest), %d words stack\r\n",
-           tskIDLE_PRIORITY, defaultTask_attributes.stack_size);
+    printf("  - Default Task: Priority %ld (lowest), %lu words stack\r\n",
+           (long)tskIDLE_PRIORITY,
+           (unsigned long)defaultTask_attributes.stack_size);
     printf("Main: Starting FreeRTOS scheduler...\r\n");
 
     /* USER CODE END RTOS_THREADS */
